Add jsonEqual to test helpers file
