+++ server/users_test.go
@@ -11,7 +11,6 @@ import (
 	"testing"
 
 	"github.com/bouk/httprouter"
-	"github.com/google/go-cmp/cmp"
 	"github.com/influxdata/chronograf"
 	"github.com/influxdata/chronograf/log"
 	"github.com/influxdata/chronograf/mocks"
@@ -525,17 +524,4 @@ func TestService_Users(t *testing.T) {
 			}
 		})
 	}
-}
-
-func jsonEqual(s1, s2 string) (eq bool, err error) {
-	var o1, o2 interface{}
-
-	if err = json.Unmarshal([]byte(s1), &o1); err != nil {
-		return
-	}
-	if err = json.Unmarshal([]byte(s2), &o2); err != nil {
-		return
-	}
-
-	return cmp.Equal(o1, o2), nil
 }
